+++ src/hardware/driver/low_timers_fake.cpp
@@ -39,7 +39,9 @@ std::chrono::high_resolution_clock::time_point low_timers_fake::start_timing()
      * @brief wait for the tick to end.
      * Remember to run start_timing first!
      */
-std::chrono::high_resolution_clock::time_point low_timers_fake::wait_for_tick_us(const std::chrono::high_resolution_clock::time_point&, const double dt)
+std::chrono::high_resolution_clock::time_point low_timers_fake::wait_for_tick_us(
+    const std::chrono::high_resolution_clock::time_point&,
+    const int64_t dt)
 {
     last_delay = dt;
     on_wait_s(dt);
